+++ rust/basic_ethereum/src/state.rs
@@ -1,0 +1,93 @@
+use crate::{EcdsaKeyName, EcdsaPublicKey, EthereumNetwork, InitArg};
+use evm_rpc_canister_types::{EthMainnetService, EthSepoliaService, RpcServices};
+use ic_cdk::api::management_canister::ecdsa::EcdsaKeyId;
+use std::cell::RefCell;
+use std::ops::{Deref, DerefMut};
+
+thread_local! {
+    pub static STATE: RefCell<State> = RefCell::default();
+}
+
+pub fn init_state(init_arg: InitArg) {
+    STATE.with(|s| *s.borrow_mut() = State::from(init_arg));
+}
+
+pub fn read_state<R>(f: impl FnOnce(&State) -> R) -> R {
+    STATE.with(|s| f(s.borrow().deref()))
+}
+
+pub fn mutate_state<F, R>(f: F) -> R
+where
+    F: FnOnce(&mut State) -> R,
+{
+    STATE.with(|s| f(s.borrow_mut().deref_mut()))
+}
+
+#[derive(Debug, Default, PartialEq, Eq)]
+pub struct State {
+    ethereum_network: EthereumNetwork,
+    ecdsa_key_name: EcdsaKeyName,
+    ecdsa_public_key: Option<EcdsaPublicKey>,
+}
+
+impl State {
+    pub fn ecdsa_key_id(&self) -> EcdsaKeyId {
+        EcdsaKeyId::from(&self.ecdsa_key_name)
+    }
+
+    pub fn ethereum_network(&self) -> EthereumNetwork {
+        self.ethereum_network
+    }
+
+    pub fn evm_rpc_services(&self) -> RpcServices {
+        match self.ethereum_network {
+            EthereumNetwork::Mainnet => RpcServices::EthMainnet(None),
+            EthereumNetwork::Sepolia => RpcServices::EthSepolia(None),
+        }
+    }
+
+    pub fn single_evm_rpc_service(&self) -> RpcServices {
+        match self.ethereum_network {
+            EthereumNetwork::Mainnet => {
+                RpcServices::EthMainnet(Some(vec![EthMainnetService::Ankr]))
+            }
+            EthereumNetwork::Sepolia => {
+                RpcServices::EthSepolia(Some(vec![EthSepoliaService::Ankr]))
+            }
+        }
+    }
+}
+
+impl From<InitArg> for State {
+    fn from(init_arg: InitArg) -> Self {
+        State {
+            ethereum_network: init_arg.ethereum_network.unwrap_or_default(),
+            ecdsa_key_name: init_arg.ecdsa_key_name.unwrap_or_default(),
+            ..Default::default()
+        }
+    }
+}
+
+pub async fn lazy_call_ecdsa_public_key() -> EcdsaPublicKey {
+    use ic_cdk::api::management_canister::ecdsa::{ecdsa_public_key, EcdsaPublicKeyArgument};
+
+    if let Some(ecdsa_pk) = read_state(|s| s.ecdsa_public_key.clone()) {
+        return ecdsa_pk;
+    }
+    let key_id = read_state(|s| s.ecdsa_key_id());
+    let (response,) = ecdsa_public_key(EcdsaPublicKeyArgument {
+        canister_id: None,
+        derivation_path: vec![],
+        key_id,
+    })
+    .await
+    .unwrap_or_else(|(error_code, message)| {
+        ic_cdk::trap(&format!(
+            "failed to get canister's public key: {} (error code = {:?})",
+            message, error_code,
+        ))
+    });
+    let pk = EcdsaPublicKey::from(response);
+    mutate_state(|s| s.ecdsa_public_key = Some(pk.clone()));
+    pk
+}
